+++ lib/background-fetch.js
@@ -17,20 +17,19 @@ if (chrome.runtime.getBackgroundPage) {
     }
     return true; // Tell browser to await response
   });
-} else {
-  // Setup in content script
-  window.backgroundFetch = function (...args) {
-    return new Promise((resolve, reject) => {
-      chrome.runtime.sendMessage({
-        action: 'fetch',
-        arguments: args
-      }, response => {
-        if (String(response).startsWith('Error')) {
-          reject(new Error(response.replace(/Error:? ?/, '')));
-        } else {
-          resolve(response);
-        }
-      });
+}
+
+export default function backgroundFetch(...args) {
+  return new Promise((resolve, reject) => {
+    chrome.runtime.sendMessage({
+      action: 'fetch',
+      arguments: args
+    }, response => {
+      if (String(response).startsWith('Error')) {
+        reject(new Error(response.replace(/Error:? ?/, '')));
+      } else {
+        resolve(response);
+      }
     });
-  };
+  });
 }
