+++ package.json
@@ -2,26 +2,42 @@
   "name": "npmhub",
   "version": "0.0.0",
   "description": "",
-  "devDependencies": {
-    "chrome-webstore-upload-cli": "^1.0.0",
-    "dot-json": "^1.0.3",
-    "webext": "^1.9.1-with-submit.1",
-    "xo": "^0.18.2"
-  },
   "private": true,
   "author": "zeke",
   "license": "MIT",
   "scripts": {
     "test": "xo",
+    "build": "rollup -c",
+    "watch": "npm run build --silent -- --watch",
     "release-amo": "cd extension && webext submit",
     "release-cws": "cd extension && webstore upload --auto-publish",
+    "release": "npm run build && npm run update-version && npm run release-amo && npm run release-cws",
     "update-version": "dot-json extension/manifest.json version $(date -u +%Y.%-m.%-d.%-H%M)"
   },
   "xo": {
     "space": 2,
+    "ignores": [
+      "extension/**"
+    ],
+    "rules": {
+      "import/no-unassigned-import": 0
+    },
     "envs": [
       "browser",
       "webextensions"
     ]
+  },
+  "devDependencies": {
+    "chrome-webstore-upload-cli": "^1.0.0",
+    "dot-json": "^1.0.3",
+    "rollup": "^0.43.0",
+    "rollup-plugin-commonjs": "^8.0.2",
+    "rollup-plugin-node-resolve": "^3.0.0",
+    "rollup-watch": "^4.0.0",
+    "webext": "^1.9.1-with-submit.1",
+    "xo": "^0.18.2"
+  },
+  "dependencies": {
+    "escape-goat": "^1.1.0"
   }
 }
