+++ src/background.js
@@ -1,0 +1,1 @@
+import '../lib/background-fetch';
